feat: 최근에 본 기기 api함수,훅,타입 구현
